move job releated test to enginer_job_test.go
